rm LibGit2 dependency
