+++ stdlib/ArtifactDownloads/Project.toml
@@ -7,7 +7,6 @@ Artifacts = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 FileWatching = "7b1f6079-737a-58dc-b8bc-7a2ca5c1b5ee"
-LibGit2 = "76f85450-5226-5b5a-8eaa-529ad045b433"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Tar = "a4e569a6-e804-4fa4-b0f3-eef7a1d5b13e"
 p7zip_jll = "3f19e933-33d8-53b3-aaab-bd5110c3b7a0"
